Revert "New keyboard shortcuts"
