+++ packages/app/app/studio/[organization]/clips/[stageId]/Timeline/index.tsx
@@ -6,6 +6,7 @@ import Playhead from './PlayHead';
 import { formatTime } from '@/lib/utils/time';
 import { useEffect, useRef } from 'react';
 import { calculateTimelineScale } from '@/lib/utils/utils';
+import { toast } from 'sonner';
 
 const Timeline = () => {
   const {
@@ -20,6 +21,12 @@ const Timeline = () => {
     pixelsPerSecond,
     setPixelsPerSecond,
     setTimelineContainerWidth,
+    currentTime,
+    setStartTime,
+    playbackStatus,
+    setEndTime,
+    startTime,
+    endTime,
     setCurrentTime,
   } = useClipContext();
 
@@ -61,6 +68,40 @@ const Timeline = () => {
       window.removeEventListener('resize', calculateScale);
     };
   }, [maxLength, setPixelsPerSecond, videoRef, timelineContainerWidth]);
+
+  // Handle keyboard shortcuts for trimming
+  useEffect(() => {
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (playbackStatus) {
+        if (event.key === 'i') {
+          if (endTime.displayTime > currentTime) {
+            setStartTime({
+              unix: Date.now() - playbackStatus.offset,
+              displayTime: currentTime,
+            });
+          } else {
+            toast.error('Start time must be less than end time');
+          }
+        } else if (event.key === 'o') {
+          if (startTime.displayTime < currentTime) {
+            setEndTime({
+              unix: Date.now() - playbackStatus.offset,
+              displayTime: currentTime,
+            });
+          } else {
+            toast.error('End time must be greater than start time');
+          }
+        }
+      }
+    };
+
+    window.addEventListener('keydown', handleKeyDown);
+
+    return () => {
+      window.removeEventListener('keydown', handleKeyDown);
+    };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [currentTime, handleMouseDown, playbackStatus, startTime, endTime]);
 
   const handleTimelineClick = (event: React.MouseEvent) => {
     if (videoRef.current) {
